+++ packages/presets/client/tests/fixtures/with-enum-values.ts
@@ -1,0 +1,5 @@
+export enum MyColor {
+  RED,
+  BLUE,
+  GREEN,
+}